fix: split server.py from main.py — BedrockAgentCoreApp has no .get()/.post()
